exported functions for readability
