--- server/routes/bus-stop.json
@@ -1,38 +1,0 @@
-{
-    "bstop" : {
-        "CW" : [
-            {"mainEntrance" : {"lat":36.982774797004936, "lon":-122.06270931737215}},
-            {"highWestern"  : {"lat":36.978804677752784, "lon":-122.0577306086418}},
-            {"arboretum"    : {"lat":36.977312373918984, "lon":-122.05431182690278}},
-            {"oakes"        : {"lat":36.990600375537326, "lon":-122.06609759888855}},
-            {"porter"       : {"lat":36.992799056682486, "lon":-122.064849623487}},
-            {"kerrBridge"   : {"lat":36.996705253649324, "lon":-122.06358662297386}},
-            {"kresge"       : {"lat":36.999205841029166, "lon":-122.06434263597127}},
-            {"scienceHill"  : {"lat":36.99985421013558,  "lon":-122.06217469848885}},
-            {"9_10"         : {"lat":36.999754697106326, "lon":-122.05831219166659}},
-            {"cowell"       : {"lat":36.99665961021545,  "lon":-122.05541186818859}},
-            {"eastLot"      : {"lat":36.991275118468785, "lon":-122.05488552653608}},
-            {"farm"         : {"lat":36.9855418147365,   "lon":-122.05355450228006}},
-            {"lowerCampus"  : {"lat":36.98155894281599,  "lon":-122.05210702172775}}
-        ],
-        "CCW" : [
-            {"mainEntance"  : {"lat":36.977638466693925, "lon":-122.0536797936101}},
-            {"lowerCampus"  : {"lat":36.98136837504213,  "lon":-122.05192995808088}},
-            {"farm"         : {"lat":36.98591498663266,  "lon":-122.05354355936608}},
-            {"eastLot"      : {"lat":36.99131463260851,  "lon":-122.05466591157932}},
-            {"eastField"    : {"lat":36.994260762954625, "lon":-122.05551548200211}},
-            {"cowell"       : {"lat":36.99749509046836,  "lon":-122.05506204353348}},
-            {"merill"       : {"lat":36.99900772270635,  "lon":-122.05517818344084}},
-            {"9_10"         : {"lat":36.999922341723256, "lon":-122.05831912968208}},
-            {"scienceHill"  : {"lat":36.99997255960926,  "lon":-122.0623325201478}},
-            {"kresge"       : {"lat":36.99928624273428,  "lon":-122.06455088687368}},
-            {"porter"       : {"lat":36.992995555190426, "lon":-122.06520358924199}},
-            {"familyHouse"  : {"lat":36.99179088764398,  "lon":-122.06680290745679}},
-            {"oakes"        : {"lat":36.98994329633098,  "lon":-122.06720247682834}},
-            {"arboretum"    : {"lat":36.98368434365876,  "lon":-122.06493451904409}},
-            {"toscaTerrace" : {"lat":36.979887992208106, "lon":-122.05933119132885}},
-            {"highWestern"  : {"lat":36.97864381292043,  "lon":-122.05782758686746}}
-        ]
-    }
-}
-
