+++ server/routes/index.js
@@ -2,7 +2,8 @@ const express = require('express')
 const router = express.Router()
 const metro = require('./metro')
 require('dotenv').config()
-const busStops = require('./bus-stops.json')
+import { calcCWorCCW } from "./direction.js"
+import { nextBusStops } from "./soonBusStop.js"
 
 // Helper functions
 const {
@@ -226,155 +227,6 @@ function dataValidate(data) {
     Object.keys(data).length != 6
   ) { return true;}
   else { return false;}
-}
-
-module.exports = router
-
-// Determine if bus is going up or down
-function latitudeDecreasing(previousLocationArray) {
-  total = 0;
-  for (let i = 0; i < previousLocationArray.length - 1; i++) {
-    if((previousLocationArray[i].lat - previousLocationArray[i+1].lat) > 0) {
-      total += 1;
-    }
-    else {
-      total -= 1;
-    }
-  }
-  if (total > 0) return false;
-  return true;
-  
-}
-
-// Determine if bus is going left or right
-function longitudeDecreasing(previousLocationArray) {
-  total = 0;
-  for (let i = 0; i < previousLocationArray.length - 1; i++) {
-    if((previousLocationArray[i].lon - previousLocationArray[i+1].lon) > 0) {
-      total += 1;
-    }
-    else {
-      total -= 1;
-    }
-  }
-  if (total > 0) return false;
-  return true;
-}
-
-// Calculate direction of bus
-function calcCWorCCW({lat1, lon1}, previousLocationArray) {
-  // Lower Half
-  if (36.977583 < lat1 && lat1 < 36.992444) {
-    // Lower West Half
-    if (lon1 < -122.055831) {
-      if (latitudeDecreasing(previousLocationArray)) return "ccw";
-      else return "cw";
-    }
-    // Lower East Half
-    else {
-      if (latitudeDecreasing(previousLocationArray)) return "cw";
-      else return "ccw";
-    }
-  }
-
-  // RCC Area
-  if ((36.992444 <= lat1 && lat1 < 36.993316) && (-122.066566 < lon1 && lon1 < -122.061)) {
-    if (longitudeDecreasing(previousLocationArray)) return "ccw";
-    else return "cw";
-  }
-
-  // RCC to Kresge
-  if ((36.993316 <= lat1 && lat1 < 36.999290) && lon1 < -122.062260) {
-    if (latitudeDecreasing(previousLocationArray)) return "ccw";
-    else return "cw";
-  }
-
-  // Baskin to Crown
-  if ((36.999290 <= lat1) && (-122.064560 <= lon1 && lon1 < -122.054543)) {
-    if (longitudeDecreasing(previousLocationArray)) return "ccw";
-    else return "cw";
-  }
-
-  // Crown to East Remote
-  if ((36.992444 <= lat1 && lat1 < 36.999290) && (lon1 >= -122.055831)) {
-    if (latitudeDecreasing(previousLocationArray)) return "cw";
-    else return "ccw";
-  }
-
-  // Bay and High Area
-  if (36.977119 < lat1 && lat1 < 36.9775833) {
-    // West Side
-    if (lon1 < -122.053795) {
-      if (longitudeDecreasing(previousLocationArray)) return "cw";
-      else return "ccw";
-    }
-    // East Side
-    if (lon1 >= -122.053795) {
-      if (latitudeDecreasing(previousLocationArray)) return "cw";
-      else return "ccw";
-    }
-  }
-  return "n/a";
-}
-
-// updates next 3 bus stops
-
-function nextBusStops({lat1, lon1}, previousLocationArray) {
-  let direction = calcCWorCCW({lat1, lon1}, previousLocationArray)
-
-  if (direction == "cw") {
-    const cwData = busStops.bstop.CW;
-    for (let i = 0; i < cwData.length; i++) {
-      let location = cwData[i];
-      let locationName = Object.keys(location)[0];
-      let lat2 = location[locationName].lat;
-      let lon2 = location[locationName].lon;
-      if((lat2 - 0.000450) <= lat1 <= (lat2 + 0.000450)) {
-        if ((lon2 - 0.000450) <= lon1 <= (lon2 + 0.000450)) {
-          let stops_arr = [Object.keys(cwData[i + 1])[0], Object.keys(cwData[i + 2])[0], Object.keys(cwData[i + 3])[0]]
-          soonBusStopUpdate(stops_arr, "CW");
-        }
-      } 
-    }
-  } 
-  else if(direction == "ccw") {
-    const ccwData = busStops.bstop.CCW;
-    for (let i = 0; i < ccwData.length; i++) {
-      let location = ccwData[i];
-      let locationName = Object.keys(location)[0];
-      let lat2 = location[locationName].lat;
-      let lon2 = location[locationName].lon;
-      if((lat2 - 0.000450) <= lat1 <= (lat2 + 0.000450)) {
-        if ((lon2 - 0.000450) <= lon1 <= (lon2 + 0.000450)) {
-          let stops_arr = [Object.keys(ccwData[i + 1])[0], Object.keys(ccwData[i + 2])[0], Object.keys(ccwData[i + 3])[0]]
-          soonBusStopUpdate(stops_arr, "CCW");
-        }
-      }
-    }
-  }
-}
-
-// Set bus stops "Soon" to true if 
-
-function soonBusStopUpdate(soonBusStops, direction) {
-  // Get a database reference to the collection of stops given a direction
-  const stopRef = defaultDatabase.collection('busStop').document(direction);
-
-  // Get all the collection values inside the doc
-  const stops =  async () => {
-    await stopRef.listCollections();
-    stops.forEach(stop => {
-      if(soonBusStops.includes(stop.id)) {
-        stop.update({
-          soon: true
-        });
-      } else {
-        stop.update({
-          soon: false
-        });
-      };
-    });
-  };
 }
 
 module.exports = router
